requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
numpy
matplotlib
seaborn
scikit-learn
python-multipart
requests
